topic2/whats-a-jwt/requirements.txt
flask==3.1.0
pyquocca==0.4.0
PyJWT==2.7.0